% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{aggregate_over_deciles}
\alias{aggregate_over_deciles}
\title{DEPRECIATED: aggregate_over_deciles()}
\usage{
aggregate_over_deciles(prepared_input)
}
\arguments{
\item{scores_and_ntiles}{Dataframe resulting from function \code{\link{prepare_scores_and_ntiles}} or a data frame
that meets requirements as specified in \code{\link{aggregate_over_deciles}} in the section: \bold{When you build scores_and_ntiles yourself} . It is made backward compatible to also work on a \code{scores_and_deciles} dataframe created in earlier version of modelplotr .}
}
\value{
Dataframe \code{ntiles_aggregate} is built based on \code{scores_and_ntiles}.\cr\cr
}
\description{
This function is depreciated, since modelplotr now supports that you can specify the number of bins (ntiles) you prefer.
You can choose a value for the \code{ntiles} parameter between 4 and 100, the default is 10 (deciles).
Therefore, please use \code{\link{aggregate_over_deciles}} instead.
}
\seealso{
use \code{\link{aggregate_over_deciles}} instead.
}
