% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{cumgains}
\alias{cumgains}
\title{Cumulative gains plot}
\usage{
cumgains(plot_input = eval_t_type)
}
\arguments{
\item{plot_input}{Dataframe. Dataframe needs to be created with input_modevalplots
or else meet required input format (see link to format in "See Also" section)}

\item{targetcat}{String.}
}
\value{
Cumulative gains plot.
}
\description{
Cumulative gains plot
}
\examples{
add(1, 1)
add(10, 10)
}
\seealso{
\code{\link{input_modevalplots}} for details on the required input dataframe format.
}
