% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{cumresponse}
\alias{cumresponse}
\title{Cumulative responnse plot}
\usage{
cumresponse(plot_input = eval_t_tot, targetcat = "setosa")
}
\arguments{
\item{plot_input}{Dataframe.}

\item{targetcat}{String.}
}
\value{
Cumulative response plot.
}
\description{
Cumulative responnse plot
}
\examples{
add(1, 1)
add(10, 10)
}
