% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{fourevalplots}
\alias{fourevalplots}
\title{Create plot with all four evaluation plots}
\usage{
fourevalplots(plot_input = eval_t_type, customlinecolors = NA)
}
\arguments{
\item{plotlist}{List of plot objects.}

\item{file}{???}

\item{cols}{Integer. Number of columns}

\item{layout}{???}
}
\value{
Multiplot constisting of plots in \code{plotlist}.
}
\description{
Create plot with all four evaluation plots
}
\examples{
add(1, 1)
add(10, 10)
}
