% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{lift}
\alias{lift}
\title{Lift plot}
\usage{
lift(plot_input = eval_t_tot, targetcat = "setosa")
}
\arguments{
\item{plot_input}{Dataframe.}

\item{targetcat}{String.}
}
\value{
Lift plot.
}
\description{
Lift plot
}
\examples{
add(1, 1)
add(10, 10)
}
