% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{multiplot}
\alias{multiplot}
\title{Create multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{List of plot objects.}

\item{file}{???}

\item{cols}{Integer. Number of columns}

\item{layout}{???}
}
\value{
Multiplot constisting of plots in \code{plotlist}.
}
\description{
Create multiplot
}
\examples{
add(1, 1)
add(10, 10)
}
