% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{plot_costsrevs}
\alias{plot_costsrevs}
\title{Create plot with all four evaluation plots}
\usage{
plot_costsrevs(data = plot_input, highlight_ntile = NA,
  highlight_how = "plot_text", save_fig = FALSE,
  save_fig_filename = NA, custom_line_colors = NA,
  custom_plot_text = NULL, fixed_costs, variable_costs_per_unit,
  profit_per_unit, ...)
}
\arguments{
\item{data}{Dataframe. Dataframe needs to be created with \code{\link{plotting_scope}}
or else meet required input format.}

\item{save_fig}{Logical. Save plot to file? Default = FALSE. When set to TRUE, saved plot_all is optimized for 36x24cm.}

\item{save_fig_filename}{String. Filename of saved plot. Default the plot is saved as {working_dir_path}/{plotname}.png.}

\item{custom_line_colors}{Vector of Strings. Specifying colors for the lines in the plot.
When not specified, colors from the RColorBrewer palet "Set1" are used.}
}
\value{
gtable, containing 6 grobs.
}
\description{
Generates a layout containing a number graphical elements, including title, subtitle and the four
model evaluation plots: cumulative gains plot, lift plot, response plot and cumulative response plot.
}
