% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{prepare_scores_and_deciles}
\alias{prepare_scores_and_deciles}
\title{DEPRECIATED: prepare_scores_and_deciles()}
\usage{
prepare_scores_and_deciles(datasets, dataset_labels, models, model_labels,
  target_column)
}
\arguments{
\item{datasets}{List of Strings. A list of the names of the dataframe
objects to include in model evaluation. All dataframes need to contain
target variable and feature variables.}

\item{dataset_labels}{List of Strings. A list of labels for the datasets, user.
When dataset_labels is not specified, the names from \code{datasets} are used.}

\item{models}{List of Strings. Names of the model objects containing parameters to
apply models to data. To use this function, model objects need to be generated
by the mlr package or by the caret package or by the h20 package.
Modelplotr automatically detects whether the model is built using mlr or caret or h2o.}

\item{model_labels}{List of Strings. Labels for the models to use in plots.
When model_labels is not specified, the names from \code{moddels} are used.}

\item{target_column}{String. Name of the target variable in datasets. Target
can be either binary or multinomial. Continuous targets are not supported.}
}
\value{
Dataframe. Dataframe \code{scores_and_ntiles} is built, based on the \code{datasets}
  and \code{models} specified. It contains the dataset name, actuals on the \code{target} ,
  the predicted probabilities for each class of the target and attribution to
  ntiles in the dataset for each class of the target.
}
\description{
This function is depreciated, since modelplotr now supports that you can specify the number of bins (ntiles) you prefer.
You can choose a value for the \code{ntiles} parameter between 4 and 100, the default is 10 (deciles).
Therefore, please use \code{\link{prepare_scores_and_ntiles}} instead.
}
\seealso{
use \code{\link{prepare_scores_and_ntiles}} instead.
}
