% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{response}
\alias{response}
\title{Response plot}
\usage{
response(plot_input = eval_t_type)
}
\arguments{
\item{plot_input}{Dataframe.}

\item{targetcat}{String.}
}
\value{
Lift plot.
}
\description{
Response plot
}
\examples{
add(1, 1)
add(10, 10)
}
