% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{savemodelplots}
\alias{savemodelplots}
\title{Save Model Plots to file.}
\usage{
savemodelplots(plots = c("cumgains", "lift", "response", "cumresponse"),
  dir = getwd())
}
\arguments{
\item{plots}{list of strings. List of plot names as strings.}

\item{dir}{String. Directory on file system where to save plots to.
Default: working directory.}
}
\value{
Save model plots to file
}
\description{
Save Model Plots to file.
}
\examples{
add(1, 1)
add(10, 10)
}
