% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{cumgains}
\alias{cumgains}
\title{Cumulative gains plot}
\usage{
cumgains(plot_input = eval_t_tot, targetcat = "setosa")
}
\arguments{
\item{plot_input}{Dataframe.}

\item{targetcat}{String.}
}
\value{
Cumulative gains plot.
}
\description{
Cumulative gains plot
}
\examples{
add(1, 1)
add(10, 10)
}
