% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{dataprep_modevalplots}
\alias{dataprep_modevalplots}
\title{Build 'eval_tot' containing Actuals and Predictions}
\usage{
dataprep_modevalplots(datasets = list("train", "test"), modelname = "clf",
  targetname = "y")
}
\arguments{
\item{datasets}{List of strings. A list of the names of the dataframe objects to include in model evaluation. All dataframes need to contain target variable and feature variables. Default value is list("train","set"), hence the presence of objects named "train" and "test" in the environment is expected.}

\item{modelname}{String. Name of the model object containing parameters to apply model on data.}

\item{targetname}{String. Name of the target variable in datasets. Target can be either binary or multinomial. Continuous targets are not supported.}
}
\value{
Dataframe. Dataframe 'eval_tot' is built based on \code{datasets}.
It contains the dataset name, actuals on the target \code{target} , the predicted probabilities for each class of the target
and attribution to deciles in the dataset for each class of the target.
}
\description{
Build dataframe object 'eval_tot' that contains actuals and predictions on dependent variable for each dataset in datasets
}
\examples{
add(1, 1)
add(10, 10)
}
\seealso{
\code{\link{input_modevalplots}} for details on the function that aggregates the output to the input for the plots.
}
