% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{fourevalplots}
\alias{fourevalplots}
\title{Create plot with all four evaluation plots}
\usage{
fourevalplots(plot_input = eval_t_type, customlinecolors = NA)
}
\arguments{
\item{plot_input}{Dataframe. Dataframe needs to be created with \code{\link{scope_modevalplots}}
or else meet required input format.}

\item{customlinecolors}{Vector of Strings. Specifying colors for the lines in the plot.
When not specified, colors from the RColorBrewer palet "Set1" are used.}
}
\value{
gtable, containing 6 grobs.
}
\description{
Generates a layout containing a number graphical elements, including title, subtitle and the four
model evaluation plots: cumulative gains plot, lift plot, response plot and cumulative response plot.
}
\examples{
data(iris)
train_index =  sample(seq(1, nrow(iris)),size = 0.7*nrow(iris), replace = F )
train = iris[train_index,]
test = iris[-train_index,]
trainTask <- mlr::makeClassifTask(data = train, target = "Species")
testTask <- mlr::makeClassifTask(data = test, target = "Species")
mlr::configureMlr() # this line is needed when using mlr without loading it (mlr::)
#estimate models
task = mlr::makeClassifTask(data = train, target = "Species")
lrn = mlr::makeLearner("classif.randomForest", predict.type = "prob")
rf = mlr::train(lrn, task)
lrn = mlr::makeLearner("classif.multinom", predict.type = "prob")
mnl = mlr::train(lrn, task)
dataprep_modevalplots(datasets=list("train","test"),
                      datasetlabels = list("train data","test data"),
                      models = list("rf","mnl"),
                      modellabels = list("random forest","multinomial logit"),
                      targetname="Species")
head(eval_tot)
input_modevalplots()
scope_modevalplots()
cumgains()
lift()
response()
cumresponse()
fourevalplots()
}
