% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{input_modevalplots}
\alias{input_modevalplots}
\title{Build 'eval_t' with aggregated evaluation measures}
\usage{
input_modevalplots(eval_tot = "eval_tot")
}
\arguments{
\item{eval_tot}{Dataframe resulting from function dataprep_modevalplots().}
}
\value{
Dataframe 'eval_t' is built based on \code{eval_t}.
It contains ..
}
\description{
Build dataframe 'eval_t' with aggregated actuals and predictions . A record in 'eval_t' is unique on the combination of datasets-decile.
}
\examples{
add(1, 1)
add(10, 10)
}
\seealso{
\code{\link{dataprep_modevalplots}} for details on the function that generated the required input.
}
