% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingmodelplots.R
\name{plot_profit}
\alias{plot_profit}
\title{Profit plot}
\usage{
plot_profit(data = plot_input, highlight_ntile = "max_profit",
  highlight_how = "plot_text", save_fig = FALSE,
  save_fig_filename = NA, custom_line_colors = NA,
  custom_plot_text = NULL, fixed_costs, variable_costs_per_unit,
  profit_per_unit, ...)
}
\arguments{
\item{data}{Dataframe. Dataframe needs to be created with \code{\link{plotting_scope}}
or else meet required input format.}

\item{highlight_ntile}{Integer or string ("max_roi" or "max_profit"). Specifying the ntile at which the plot is annotated
and/or performances are highlighted. Default value is \code{max_profit}, highlighting the ntile where profit is highest.}

\item{highlight_how}{String. How to annotate the plot. Possible values: "plot_text","plot", "text".
Default is "plot_text", both highlighting the ntile and value on the plot as well as in text below the plot.
"plot" only highligths the plot, but does not add text below the plot explaining the plot at chosen ntile.
"text" adds text below the plot explaining the plot at chosen ntile but does not highlight the plot.}

\item{save_fig}{Logical. Save plot to file? Default = FALSE. When set to TRUE, saved plot is optimized for 36x24cm.}

\item{save_fig_filename}{String. Filename of saved plot. Default the plot is saved as {working_dir_path}/{plotname}.png.}

\item{custom_line_colors}{Vector of Strings. Specifying colors for the lines in the plot.
When not specified, colors from the RColorBrewer palet "Set1" are used.}

\item{custom_plot_text}{List. List with customized textual elements for plot. Create a list with defaults
by using \code{\link{customize_plot_text}} and override default values to customize.}

\item{fixed_costs}{Numeric. Specifying the fixed costs related to a selection based on the model.
These costs are constant and do not vary with selection size (ntiles).}

\item{variable_costs_per_unit}{Numeric. Specifying the variable costs per selected unit for a selection based on the model.
These costs vary with selection size (ntiles).}

\item{profit_per_unit}{Numeric. Specifying the profit per unit in case the selected unit converts / responds positively.}
}
\value{
gtable, containing 6 grobs.
}
\description{
Generates the Profit plot. It plots the cumulative profit up until that ntile when the model is used for campaign selection.
It can be used to answer the following business question:
When we apply the model and select up until ntile X, what is the expected profit of the campaign?
Extra parameters needed for this plot are: fixed_costs, variable_costs_per_unit and profit_per_unit.
}
