% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{plotting_scope}
\alias{plotting_scope}
\title{Build 'plot_input' with subset for selected evaluation type.}
\usage{
plotting_scope(prepared_input = deciles_aggregate, scope = "no_comparison",
  select_model_label = NA, select_dataset_label = NA,
  select_targetclass = NA, select_smallest_targetclass = TRUE)
}
\arguments{
\item{prepared_input}{Dataframe. Dataframe resulting from function \code{\link{aggregate_over_deciles}()} or with similar
format. Default value is deciles_aggregate, the output of \code{aggregate_over_deciles()}. When deciles_aggregate is not found, function
\code{aggregate_over_deciles()} is automatically called.}

\item{scope}{String. Evaluation type of interest. Possible values:
"compare_models","compare_datasets", "compare_targetclasses","no_comparison".
Default is NA, equivalent to "no_comparison".}

\item{select_model_label}{String. Selected model when scope is "compare_datasets" or "compare_targetclasses" or "no_comparison".
Needs to be identical to model descriptions as specified in model_labels (or models when model_labels is not specified).
When scope is "compare_models", select_model_label can be used to take a subset of available models.}

\item{select_dataset_label}{String. Selected dataset when scope is compare_models or compare_targetclasses or no_comparison.
Needs to be identical to dataset descriptions as specified in dataset_labels (or datasets when dataset_labels is not
specified). When scope is "compare_datasets", select_dataset_label can be used to take a subset of available datasets.}

\item{select_targetclass}{String. Selected target value when scope is compare_models or compare_datasets or no_comparison.
Default is smallest value when select_smallest_targetclass=TRUE, otherwise first alphabetical value.
When scope is "compare_targetclasses", select_targetclass can be used to take a subset of available target classes.}

\item{select_smallest_targetclass}{Boolean. Select the target value with the smallest number of cases in dataset as group of
interest. Default is True, hence the target value with the least observations is selected.}
}
\value{
Dataframe \code{plot_input} is a subset of \code{deciles_aggregate}.
}
\description{
Build dataframe 'plot_input' with a subset of 'deciles_aggregate' that meets the requested evaluation perspective.
There are four perspectives:
\describe{
  \item{"no_comparison" (default)}{In this perspective, you're interested in the performance of one model on one dataset
    for one target class. Therefore, only one line is plotted in the plots.
    The parameters \code{select_model_label}, \code{select_dataset_label} and \code{select_targetclass} determine which group is
    plotted. When not specified, the first alphabetic model, the first alphabetic dataset and
    the smallest (when \code{select_smallest_targetclass=TRUE}) or first alphabetic target value are selected }
  \item{"compare_models"}{In this perspective, you're interested in how well different models perform in comparison to
    each other on the same dataset and for the same target value. This results in a comparison between models available
    in deciles_aggregate$model_label for a selected dataset (default: first alphabetic dataset) and for a selected target value
    (default: smallest (when \code{select_smallest_targetclass=TRUE}) or first alphabetic target value).}
  \item{"compare_datasets"}{In this perspective, you're interested in how well a model performs in different datasets
  for a specific model on the same target value. This results in a comparison between datasets available in
  deciles_aggregate$dataset_label for a selected model (default: first alphabetic model) and for a selected target value (default:
  smallest (when \code{select_smallest_targetclass=TRUE}) or first alphabetic target value).}
  \item{"compare_targetclasses"}{In this perspective, you're interested in how well a model performs for different target
   values on a specific dataset.This resuls in a comparison between target classes available in deciles_aggregate$target_class for
   a selected model (default: first alphabetic model) and for a selected dataset (default: first alphabetic dataset).}}
}
\examples{
data(iris)
# add some noise to iris to prevent perfect models
addNoise <- function(x) round(rnorm(n=100,mean=mean(x),sd=sd(x)),1)
iris_addnoise <- as.data.frame(lapply(iris[1:4], addNoise))
iris_addnoise$Species <- sample(unique(iris$Species),100,replace=TRUE)
iris <- rbind(iris,iris_addnoise)
train_index =  sample(seq(1, nrow(iris)),size = 0.7*nrow(iris), replace = F )
train = iris[train_index,]
test = iris[-train_index,]
trainTask <- mlr::makeClassifTask(data = train, target = "Species")
testTask <- mlr::makeClassifTask(data = test, target = "Species")
mlr::configureMlr() # this line is needed when using mlr without loading it (mlr::)
# estimate models
task = mlr::makeClassifTask(data = train, target = "Species")
lrn = mlr::makeLearner("classif.randomForest", predict.type = "prob")
rf = mlr::train(lrn, task)
lrn = mlr::makeLearner("classif.multinom", predict.type = "prob")
mnl = mlr::train(lrn, task)
prepare_scores_and_deciles(datasets=list("train","test"),
                      dataset_labels = list("train data","test data"),
                      models = list("rf","mnl"),
                      model_labels = list("random forest","multinomial logit"),
                      target_column="Species")
# preparation steps
head(scores_and_deciles)
aggregate_over_deciles()
plotting_scope()
# various plotting examples with different plotting scopes
plot_cumgains()
plot_cumgains(highlight_decile=2)
plotting_scope(scope="compare_models")
plot_cumlift()
plot_cumlift(highlight_decile=2,highlight_how="plot")
plotting_scope(scope="compare_targetclasses")
plot_response()
plot_response(custom_line_colors = c('green','orange','darkblue'))
plotting_scope(scope="compare_datasets")
plot_cumresponse()
plot_cumresponse(highlight_decile=2,highlight_how="text")
plot_all()
}
\seealso{
\code{\link{modelplotr}} for generic info on the package \code{moddelplotr}

\code{\link{aggregate_over_deciles}} for details on the function \code{aggregate_over_deciles} that
generates the required input.

\code{\link{prepare_scores_and_deciles}} for details on the function \code{prepare_scores_and_deciles}
that generates the required input.
filters the output of \code{aggregate_over_deciles} to prepare it for the required evaluation.

\url{https://github.com/modelplot/modelplotr} for details on the package

\url{https://modelplot.github.io/} for our blog on the value of the model plots
}
