% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{scope_modevalplots}
\alias{scope_modevalplots}
\title{Build 'eval_t_type' with subset for selected evaluation type.}
\usage{
scope_modevalplots(prepared_input = eval_t_tot, eval_type = "NoComparison",
  select_model = NA, select_dataset = NA, select_targetvalue = NA,
  select_smallesttargetvalue = TRUE)
}
\arguments{
\item{prepared_input}{Dataframe. Dataframe resulting from function dataprep_modevalplots() or with similar output format.}

\item{eval_type}{String. Evaluation type of interest. Possible values:
"CompareModels","CompareDatasets", "CompareTargetValues","NoComparison".
Default is NA, equivalent to "NoComparison".}

\item{select_model}{String. Selected model when eval_type is CompareDatasets or CompareTargetValues or NoComparison.
Equivalent to model descriptions as specified in modellabels (or models when modellabels is not specified).}

\item{select_dataset}{String. Selected dataset when eval_type is CompareModels or CompareTargetValues or NoComparison.
Equivalent to dataset descriptions as specified in datasetlabels (or datasets when datasetlabels is not specified).}

\item{select_targetvalue}{String. Selected target value when eval_type is CompareModels or CompareDatasets or NoComparison. Default is smallest value when select_smallesttargetvalue=TRUE, otherwise first alphabetical value.}

\item{select_smallesttargetvalue}{Boolean. Select the target value with the smallest number of cases in dataset as group of interest. Default is True, hence the smallest group is selected.}
}
\value{
Dataframe \code{eval_t_type} is a subset of \code{eval_t_tot}.
}
\description{
Build dataframe 'eval_t_type' with a subset of 'eval_t_tot' that meets the requested evaluation perspective.
There are four perspectives:
\describe{
  \item{"NoComparison"}{(default) In this perspective, only one line is plotted in the plots.
    Parameters \code{select_model}, \code{select_dataset} and \code{select_targetvalue} determine which group is
    plotted. When not specified, the first alphabetic model, the first alphabetic dataset and
    the smallest (when \code{select_smallesttargetvalue=TRUE}) or first alphabetic target value are selected }
  \item{"CompareModels"}{Comparison between models available in eval_t_tot$modelname for selected dataset
  (default: first alphabetic dataset) and for selected target value (default: smallest).}
  \item{"CompareDatasets"}{Comparison between datasets available in eval_t_tot$dataset for selected model
  (default: first alphabetic model) and for selected target value (default: smallest).}
  \item{"CompareTargetValues"}{Comparison between target values available in eval_t_tot$category for selected model
  (default: first alphabetic model) and for selected dataset (default: first alphabetic dataset).}}
}
\examples{
data(iris)
train_index =  sample(seq(1, nrow(iris)),size = 0.7*nrow(iris), replace = F )
train = iris[train_index,]
test = iris[-train_index,]
trainTask <- mlr::makeClassifTask(data = train, target = "Species")
testTask <- mlr::makeClassifTask(data = test, target = "Species")
mlr::configureMlr() # this line is needed when using mlr without loading it (mlr::)
#estimate models
task = mlr::makeClassifTask(data = train, target = "Species")
lrn = mlr::makeLearner("classif.randomForest", predict.type = "prob")
rf = mlr::train(lrn, task)
lrn = mlr::makeLearner("classif.multinom", predict.type = "prob")
mnl = mlr::train(lrn, task)
dataprep_modevalplots(datasets=list("train","test"),
                      datasetlabels = list("train data","test data"),
                      models = list("rf","mnl"),
                      modellabels = list("random forest","multinomial logit"),
                      targetname="Species")
head(eval_tot)
input_modevalplots()
scope_modevalplots()
cumgains()
lift()
response()
cumresponse()
multiplot(cumgains(),lift(),response(),cumresponse(),cols=2)
}
\seealso{
\code{\link{modelplotr}} for generic info on the package \code{moddelplotr}

\code{\link{input_modevalplots}} for details on the function \code{input_modevalplots} that
generates the required input.

\code{\link{dataprep_modevalplots}} for details on the function \code{dataprep_modevalplots}
that generates the required input.
filters the output of \code{input_modevalplots} to prepare it for the required evaluation.

\url{https://github.com/jurrr/modelplotr} for details on the package

\url{https://cmotions.nl/publicaties/} for our blog on the value of the model plots
}
