% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepmodelplots.R
\name{scope_modevalplots}
\alias{scope_modevalplots}
\title{Build 'eval_t_type' with subset for selected evaluation type.}
\usage{
scope_modevalplots(prepared_input = eval_t_tot,
  eval_type = c("CompareTrainTest", "CompareModels", "TargetValues"),
  select_model, select_dataset, select_targetvalue)
}
\arguments{
\item{eval_t_tot}{Dataframe resulting from function dataprep_modevalplots().}
}
\value{
Dataframe \code{eval_t_type} is a subset of \code{eval_t_tot}.
}
\description{
Build dataframe 'eval_t_type' subset for selected evaluation type.
There are three perspectives to take:
.
}
\examples{
add(1, 1)
add(10, 10)
}
\seealso{
\code{\link{input_modevalplots}} for details on the function that generated the required input.
}
